ensure bvec reader is tested on actual data
